only pagination left

+++ app/Http/Controllers/PostController.php
@@ -9,26 +9,36 @@ use Illuminate\Support\Facades\Session;
 class PostController extends Controller
 {
     //
-    public function index(){
-        $posts = auth()->user()->posts;
-        return view('posts.index',compact('posts'));
+    public function index()
+    {
+//        $posts = auth()->user()->posts;
+        $posts = Post::all();
+        return view('posts.index', compact('posts'));
     }
-    public function show (Post $post){
 
-      //the post_id up there must have the same name as the variable that was used to hold the post id in the
+    public function show(Post $post)
+    {
+
+        //the post_id up there must have the same name as the variable that was used to hold the post id in the
         //route file.
-        return view('layouts.blog-post',compact('post'));
+        return view('layouts.blog-post', compact('post'));
     }
-    public function edit(Post $post){
 
-        return view('posts.edit',compact('post'));
+    public function edit(Post $post)
+    {
+
+        return view('posts.edit', compact('post'));
     }
-    public function create(){
+
+    public function create()
+    {
+
 
         return view('posts.create');
     }
 
-    public function update(Post $post){
+    public function update(Post $post)
+    {
         //use the first method here when there's no image or video handling carried out separately
 //        $post->update( Request()->validate([
 //            'title'=>'bail|required|min:8',
@@ -41,42 +51,47 @@ class PostController extends Controller
 //        $post->title = request('title');
 //        auth()->user()->posts()->save($post);
         $inputs = Request()->validate([
-            'title'=>'bail|required|unique:posts|min:8',
-            'post_image'=> 'mimes:jpg,bmp,png',
-            'body'=>'bail|required|min:30'
+            'title' => 'bail|required|unique:posts|min:8',
+            'post_image' => 'mimes:jpg,bmp,png',
+            'body' => 'bail|required|min:30'
         ]);
         $post->title = $inputs['title'];
         $post->body = $inputs['body'];
 
 
-        if(Request('post_image')){
+        if (Request('post_image')) {
             $inputs['post_image'] = Request('post_image')->store('images');
             $post->post_image = $inputs['post_image'];
+
         }
+        $this->authorize('update',$post);
         $post->update();
 
         return redirect()->route('post.index');
     }
+
     public function store(Request $request)
     {
         $inputs = Request()->validate([
-            'title'=>'bail|required|unique:posts|min:8',
-            'post_image'=> 'mimes:jpg,bmp,png',
-            'body'=>'bail|required|min:30'
+            'title' => 'bail|required|unique:posts|min:8',
+            'post_image' => 'mimes:jpg,bmp,png',
+            'body' => 'bail|required|min:30'
         ]);
-        if(Request('post_image')){
+        if (Request('post_image')) {
             $inputs['post_image'] = Request('post_image')->store('images');
 
         }
+        $this->authorize('create',Post::class);
         auth()->user()->posts()->create($inputs);
-        Session::flash('creation_message','A new post was just created.');
+        Session::flash('creation_message', 'A new post was just created.');
         return redirect()->route('post.index');
     }
 
-    public function destroy(Post $post){
-
+    public function destroy(Post $post)
+    {
+        $this->authorize('delete',$post);
         $post->delete();
-        Session::flash('message','The post with title '.$post->title.'was deleted.');
+        Session::flash('message', 'The post with title ' . $post->title . 'was deleted.');
         return back();
 
     }
